# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tool Control Plane API
  version: 0.1.0
  description: Governed tool gateway for AI agents with policy, approvals, audit, and provider adapters.
servers:
  - url: http://localhost:4100
security:
  - bearerAuth: []
paths:
...
  /v1/tool-calls:
    get:
      summary: List stored tool call records
      operationId: listToolCalls
      responses:
        "200":
          description: Ordered tool call records with redacted arguments and results.
          content:
            application/json:
              schema:
                type: object
                required: [tool_calls]
                properties:
                  tool_calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/ToolCallRecord"
    post:
      summary: Execute or request approval for a governed tool call
      operationId: callTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolCallRequest"
      responses:
        "200":
          description: Tool call result, policy denial, validation error, or approval request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallResponse"
        "400":
          description: Invalid JSON request body.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
    JSONValue:
      description: Arbitrary JSON value.
    JSONObject:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/JSONValue"
...
    ToolCallRequest:
      type: object
      required: [org_id, actor_user_id, agent_run_id, service_id, environment, capability, action]
      properties:
        org_id:
          type: string
        request_id:
          type: string
        actor_user_id:
          type: string
        agent_run_id:
          type: string
        service_id:
          type: string
        environment:
          type: string
        capability:
          type: string
        action:
          type: string
        arguments:
          $ref: "#/components/schemas/JSONObject"
    ToolCallResponse:
      type: object
      required: [status, risk_level]
      properties:
        status:
          type: string
          enum: [success, denied, invalid, approval_required, error]
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        provider:
          type: string
        result:
          $ref: "#/components/schemas/JSONObject"
        reason:
          type: string
        error:
          $ref: "#/components/schemas/ToolCallError"
        approval_required:
          type: boolean
        approval_request_id:
          type: string
    ToolCallError:
      type: object
      required: [retryable]
      properties:
        provider:
          type: string
        category:
          type: string
        operation:
          type: string
        status_code:
          type: integer
        attempts:
          type: integer
        retryable:
          type: boolean
        message:
          type: string
    ToolCallRecord:
      type: object
      required:
        - id
        - at
        - org_id
        - actor_user_id
        - agent_run_id
        - service_id
        - environment
        - capability
        - action
        - risk_level
        - decision
        - status
      properties:
        id:
          type: string
          example: tool_call_000001
        at:
          type: string
          format: date-time
        request_id:
          type: string
        org_id:
          type: string
        actor_user_id:
          type: string
        agent_run_id:
          type: string
        service_id:
          type: string
        environment:
          type: string
        capability:
          type: string
        action:
          type: string
        arguments:
          $ref: "#/components/schemas/JSONObject"
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        decision:
          type: string
          enum: [allowed, denied, invalid, approval_required, approved_executed]
        provider:
          type: string
        status:
          type: string
          enum: [success, denied, invalid, approval_required, error]
        reason:
          type: string
        error:
          $ref: "#/components/schemas/ToolCallError"
        approval_request_id:
          type: string
        result:
          $ref: "#/components/schemas/JSONObject"
...
    RiskLevel:
      type: string
      enum: [unknown, read, write_low, write_high]